removed unintentional change to SummaryReportGenerator

+++ BridgeCareApp/BridgeCareCore/Services/SummaryReport/SummaryReportGenerator.cs
@@ -99,12 +99,12 @@ namespace BridgeCareCore.Services.SummaryReport
             // sorting the sections based on facility name. This is helpful throughout the report
             // generation process
             reportOutputData.InitialSectionSummaries.Sort(
-                    (a, b) => Int64.Parse(a.FacilityName.Split('-')[0]).CompareTo(Int64.Parse(b.FacilityName.Split('-')[0]))
+                    (a, b) => int.Parse(a.FacilityName.Split('-')[0]).CompareTo(int.Parse(b.FacilityName.Split('-')[0]))
                     );
             foreach (var yearlySectionData in reportOutputData.Years)
             {
                 yearlySectionData.Sections.Sort(
-                    (a, b) => Int64.Parse(a.FacilityName.Split('-')[0]).CompareTo(Int64.Parse(b.FacilityName.Split('-')[0]))
+                    (a, b) => int.Parse(a.FacilityName.Split('-')[0]).CompareTo(int.Parse(b.FacilityName.Split('-')[0]))
                     );
             }
